DNN_light_TTower/requirements.txt
torch
pandas
numpy
scikit-learn
psycopg2-binary
